# docs/api/openapi.yml
openapi: "3.1.0"
info:
  title: openNBA BFF API
  version: "1.0.0"
  description: >
    Next.js BFF API for the openNBA Next Best Action platform.
    All endpoints require a valid JWT session cookie (set by Auth.js).
    Rate limits: 5 req/min on POST /api/v1/agent/run.
servers:
  - url: https://opennba.vercel.app
    description: Production
  - url: http://localhost:3000
    description: Local development
security:
  - sessionCookie: []
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: next-auth.session-token
  schemas:
    Error:
      type: object
      required: [error, code]
      properties:
        error:
          type: string
        code:
          type: string
        details:
          type: object
...
paths:
...
  /api/v1/agent/status/{run_id}:
    get:
      summary: Poll pipeline run status
      operationId: getAgentRunStatus
      tags: [Agent]
      parameters:
        - name: run_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Run status
          content:
            application/json:
              schema:
                type: object
                properties:
                  runId:
                    type: string
                    format: uuid
                  status:
                    type: string
                    enum: [QUEUED, RUNNING, COMPLETED, FAILED, TIMEOUT, PARTIALLY_ENRICHED]
                  cardsGenerated:
                    type: integer
                    nullable: true
                  completedAt:
                    type: string
                    format: date-time
                    nullable: true
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Run not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"